gallivm: fixup const int64 builder.
Pointed out by Ilia.

Fixes: 84ba00877496 (gallivm: add 64-bit const int creator.)
Acked-by: Alyssa Rosenzweig <alyssa.rosenzweig@collabora.com>

+++ src/gallium/auxiliary/gallivm/lp_bld_const.h
@@ -127,7 +127,7 @@ lp_build_const_int32(struct gallivm_state *gallivm, int i)
 }
 
 static inline LLVMValueRef
-lp_build_const_int64(struct gallivm_state *gallivm, int i)
+lp_build_const_int64(struct gallivm_state *gallivm, int64_t i)
 {
    return LLVMConstInt(LLVMInt64TypeInContext(gallivm->context), i, 0);
 }
